+++ libft/ft_printf/manage_str.c
@@ -35,7 +35,7 @@ ssize_t	ft_printf_manage_wchar(char **format, va_list *args, t_data *data)
 	if (data->got_width && data->right_pad)
 		ft_printf_width_pad(len, data->width, data->zero_pad ? '0' : ' ',
 				data->fd);
-	return (data->got_width ? ft_max(len, data->width) : len);
+	return (data->got_width ? ft_max(len, data->width) : (ssize_t)len);
 }
 
 ssize_t	ft_printf_manage_char(char **format, va_list *args, t_data *data)
@@ -75,6 +75,6 @@ ssize_t	ft_printf_manage_str(char **format, va_list *args, t_data *data)
 		write(data->fd, str, strlen);
 		if (data->got_width && data->right_pad)
 			ft_printf_width_pad(strlen, data->width, ' ', data->fd);
-		return (data->got_width ? ft_max(strlen, data->width) : strlen);
+		return (data->got_width ? ft_max(strlen, data->width) : (ssize_t)strlen);
 	}
 }
